Some apperance improvements

+++ source/coord_system.cpp
@@ -3,8 +3,8 @@
 //--------------------------------------------------------------------------------------------------------------------
 
 Coord_System::Coord_System(double x0, double y0, double base_len, sf::Color color) :
-	x0 (x0),
-	y0 (y0),
+	x0       (x0),
+	y0       (y0),
 	base_len (base_len),
 	color    (color)
 	{}
@@ -22,27 +22,27 @@ Coord_System::~Coord_System()
 void DrawCircles(sf::RenderWindow* window, Coord_System* cs)
 {
 	sf::CircleShape circle(RADIUS);
-    circle.setFillColor(cs->color);
+	circle.setFillColor(cs->color);
 
-	for (float x = cs->x0; (x <= cs->x0 + LEN_OF_AXIS - RADIUS) and (x <= WIDTH_OF_WINDOW - RADIUS);  x += cs->base_len)
+	for (double x = cs->x0; (x <= cs->x0 + LEN_OF_AXIS - RADIUS) and (x <= WIDTH_OF_WINDOW - RADIUS);  x += cs->base_len)
 	{
 		circle.setPosition(sf::Vector2f(x - RADIUS, cs->y0 - RADIUS));
 		window->draw(circle);
 	}
 	
-	for (float y = cs->y0; (y <= cs->y0 + LEN_OF_AXIS - RADIUS) and (y <= HEIGHT_OF_WINDOW - RADIUS); y += cs->base_len)
+	for (double y = cs->y0; (y <= cs->y0 + LEN_OF_AXIS - RADIUS) and (y <= HEIGHT_OF_WINDOW - RADIUS); y += cs->base_len)
 	{
 		circle.setPosition(sf::Vector2f(cs->x0 - RADIUS, y - RADIUS));
 		window->draw(circle);
 	}
 
-	for (float x = cs->x0; (x >= cs->x0 - LEN_OF_AXIS + RADIUS) and (x >= RADIUS); x -= cs->base_len)
+	for (double x = cs->x0; (x >= cs->x0 - LEN_OF_AXIS + RADIUS) and (x >= RADIUS); x -= cs->base_len)
 	{
 		circle.setPosition(sf::Vector2f(x - RADIUS, cs->y0 - RADIUS));
 		window->draw(circle);
 	}
 
-	for (float y = cs->y0; (y >= cs->y0 - LEN_OF_AXIS + RADIUS) and (y >= RADIUS); y -= cs->base_len)
+	for (double y = cs->y0; (y >= cs->y0 - LEN_OF_AXIS + RADIUS) and (y >= RADIUS); y -= cs->base_len)
 	{
 		circle.setPosition(sf::Vector2f(cs->x0 - RADIUS, y - RADIUS));
 		window->draw(circle);
@@ -59,14 +59,12 @@ void Coord_System::draw(sf::RenderWindow* window)
 	if (Coord_System::x0 + LEN_OF_AXIS < WIDTH_OF_WINDOW) x_right = Coord_System::x0 + LEN_OF_AXIS;
 	else                                                  x_right = WIDTH_OF_WINDOW;        
 
-	if (Coord_System::x0 - LEN_OF_AXIS > 0) x_left = Coord_System::x0 - LEN_OF_AXIS;
-	else                                    x_left = 0;
+	if (Coord_System::x0 - LEN_OF_AXIS > 0)               x_left  = Coord_System::x0 - LEN_OF_AXIS;
+	else                                                  x_left  = 0;
 
 	sf::Vertex x_axis[] =
-	{
-		sf::Vertex(sf::Vector2f(x_left,  Coord_System::y0), Coord_System::color),
-		sf::Vertex(sf::Vector2f(x_right, Coord_System::y0), Coord_System::color)
-	};
+		{ sf::Vertex(sf::Vector2f(x_left,  Coord_System::y0), Coord_System::color),
+		  sf::Vertex(sf::Vector2f(x_right, Coord_System::y0), Coord_System::color) };
 	window->draw(x_axis, 2, sf::Lines);
 
 
@@ -76,14 +74,12 @@ void Coord_System::draw(sf::RenderWindow* window)
 	if (Coord_System::y0 + LEN_OF_AXIS < HEIGHT_OF_WINDOW) y_down = Coord_System::y0 + LEN_OF_AXIS;
 	else                                                   y_down = HEIGHT_OF_WINDOW;
 
-	if (Coord_System::y0 - LEN_OF_AXIS > 0)   y_up = Coord_System::y0 - LEN_OF_AXIS;
-	else                                      y_up = 0;
+	if (Coord_System::y0 - LEN_OF_AXIS > 0)                y_up   = Coord_System::y0 - LEN_OF_AXIS;
+	else                                                   y_up   = 0;
 
 	sf::Vertex y_axis[] =
-	{
-		sf::Vertex(sf::Vector2f(Coord_System::x0, y_down),  Coord_System::color),
-		sf::Vertex(sf::Vector2f(Coord_System::x0, y_up),    Coord_System::color)
-	};
+		{ sf::Vertex(sf::Vector2f(Coord_System::x0, y_down),  Coord_System::color),
+		  sf::Vertex(sf::Vector2f(Coord_System::x0, y_up),    Coord_System::color) };
 	window->draw(y_axis, 2, sf::Lines);
